Reorganize and move a few traits to new `ops` sub-module
