+++ src/ops.rs
@@ -1,0 +1,10 @@
+pub mod dot;
+pub mod exp;
+pub mod norm;
+pub mod pow;
+
+pub use dot::Dot;
+pub use exp::Exp;
+pub use norm::Norm;
+pub use pow::Pow;
+
